fixes: cast issues
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/soma/soma-io/trunk@76607 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/libcartobase/type/voxelhsv.h
@@ -466,7 +466,7 @@ namespace carto {
       std::cout << "HSV:: HSV + float" << std::endl;
     #endif
     VoxelHSV result( aa );
-    return result += bb;
+    return result += static_cast<uint8_t>(bb);
   }
  
   inline
@@ -476,7 +476,7 @@ namespace carto {
       std::cout << "HSV:: HSV + double" << std::endl;
     #endif
     VoxelHSV result( aa );
-    return result += bb;
+    return result += static_cast<uint8_t>(bb);
   }
   
   inline
@@ -525,7 +525,7 @@ namespace carto {
     #ifdef CARTO_DEBUG_RGB
       std::cout << "HSV:: float + HSV" << std::endl;
     #endif
-    VoxelHSV result( aa );
+    VoxelHSV result( static_cast<uint8_t>(aa) );
     return result += bb;
   }
   
@@ -535,7 +535,7 @@ namespace carto {
     #ifdef CARTO_DEBUG_RGB
       std::cout << "HSV:: double + HSV" << std::endl;
     #endif
-    VoxelHSV result( aa );
+    VoxelHSV result( static_cast<uint8_t>(aa) );
     return result += bb;
   }
 
@@ -596,7 +596,7 @@ namespace carto {
       std::cout << "HSV:: HSV - float" << std::endl;
     #endif
     VoxelHSV result( aa );
-    return result -= bb;
+    return result -= static_cast<uint8_t>(bb);
   }
  
   inline
@@ -606,7 +606,7 @@ namespace carto {
       std::cout << "HSV:: HSV - double" << std::endl;
     #endif
     VoxelHSV result( aa );
-    return result -= bb;
+    return result -= static_cast<uint8_t>(bb);
   }
   
   inline
@@ -655,7 +655,7 @@ namespace carto {
     #ifdef CARTO_DEBUG_RGB
       std::cout << "HSV:: float - HSV" << std::endl;
     #endif
-    VoxelHSV result( aa );
+    VoxelHSV result( static_cast<uint8_t>(aa) );
     return result -= bb;
   }
   
@@ -665,7 +665,7 @@ namespace carto {
     #ifdef CARTO_DEBUG_RGB
       std::cout << "HSV:: double - HSV" << std::endl;
     #endif
-    VoxelHSV result( aa );
+    VoxelHSV result( static_cast<uint8_t>(aa) );
     return result -= bb;
   }
